requirements.txt
fastapi==0.97.0
pydantic==1.10.9
uvicorn==0.22.0
python-multipart==0.0.6
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cpu
torchvision==0.15.2+cpu
Pillow==9.5.0